06-r-and-sql.html: myDB -> DB at line 137
In this lesson there are two databases called `myDB`. One goes with connection `conn` and the other with `myConn`. I had some difficulty understanding clearly that the `myDB`s led to separate operations. I recommend naming the first `DB` to match with the `conn` and leave the `myDB` to match with the `myConn`. There might also need to be a change to the `.RMD` file.